Use variable of type long to store a style flags returned by GetWindowStyleFlag function.
Since wxWindowBase::GetWindowStyleFlag() returns long the variable used to hold returned value should be also of the same type.

+++ src/propgrid/manager.cpp
@@ -690,7 +690,7 @@ void wxPropertyGridManager::DoThaw()
 
 void wxPropertyGridManager::SetWindowStyleFlag( long style )
 {
-    int oldWindowStyle = GetWindowStyleFlag();
+    long oldWindowStyle = GetWindowStyleFlag();
 
     wxWindow::SetWindowStyleFlag( style );
     m_pPropGrid->SetWindowStyleFlag( (m_pPropGrid->GetWindowStyleFlag()&~(wxPG_MAN_PASS_FLAGS_MASK)) |
